requirements.txt
telethon==1.36.0
python-dotenv==1.0.1
pydantic==2.9.2
asyncpg==0.29.0
loguru==0.7.2
